decouple IntegersTest and BizarroIntegersTest
