+++ src/test/java/com/esaulpaugh/headlong/util/Utils.java
@@ -1,5 +1,9 @@
 package com.esaulpaugh.headlong.util;
 
+import com.esaulpaugh.headlong.util.exception.DecodeException;
+
+import java.util.concurrent.RecursiveAction;
+
 public class Utils {
 
     public static int insertBytes(int n, byte[] b, int i, byte w, byte x, byte y, byte z) {
@@ -36,6 +40,77 @@ public class Utils {
         case 1: b[i] = z;
         case 0: return n;
         default: throw new IllegalArgumentException("n is out of range: " + n);
+        }
+    }
+
+    public static class IntTask extends RecursiveAction {
+
+        private static final int THRESHOLD = 250_000_000;
+
+        protected final long start, end;
+
+        public IntTask(long start, long end) {
+            this.start = start;
+            this.end = end;
+        }
+
+        @Override
+        protected void compute() {
+            final long n = end - start;
+            if (n > THRESHOLD) {
+                long midpoint = start + (n / 2);
+                invokeAll(
+                        new IntTask(start, midpoint),
+                        new IntTask(midpoint, end)
+                );
+            } else {
+                doWork();
+            }
+        }
+
+        protected void doWork() {
+            byte[] four = new byte[4];
+            try {
+                final long end = this.end;
+                for (long lo = this.start; lo <= end; lo++) {
+                    int i = (int) lo;
+                    int len = Integers.putInt(i, four, 0);
+                    int r = Integers.getInt(four, 0, len);
+                    if(i != r) {
+                        throw new AssertionError(i + " !=" + r);
+                    }
+                }
+            } catch (DecodeException e) {
+                throw new RuntimeException(e);
+            }
+        }
+    }
+
+    public static class LenIntTask extends IntTask {
+
+        public LenIntTask(long start, long end) {
+            super(start, end);
+        }
+
+        protected int len(int val) {
+            return Integers.len(val);
+        }
+
+        @Override
+        protected void doWork() {
+            final long end = this.end;
+            for (long lo = this.start; lo <= end; lo++) {
+                int i = (int) lo;
+                int expectedLen = i < 0 || i >= 16_777_216 ? 4
+                        : i >= 65_536 ? 3
+                        : i >= 256 ? 2
+                        : i != 0 ? 1
+                        : 0;
+                int len = LenIntTask.this.len(i); // len(int) can be overridden by subclasses
+                if(expectedLen != len) {
+                    throw new AssertionError(expectedLen + " != " + len);
+                }
+            }
         }
     }
 }
